[windows_fonts] Be sure to check the updated fonts
Fix #14
In some circumstances, in the method _get_fonts_filename_windows_10_or_more, the IDWriteFontSet doesn't contain the latest system fonts. It seems to only be the case when the font is installed with FontBase. I don't know which Win API FontBase uses to activate the font, so I haven't been able to reproduce the problem without FontBase. Also, I have only been able to reproduce this issue on Windows 11, but it may also occur on any other Windows version. Because of that, as a precaution, let's also modify the _get_fonts_filename_windows_vista_sp2_or_more method so it also checks for updated fonts