[windows_fonts] Ignore simulated font to get better performance
We don't care about the simulated font. It just create many duplicate fonts. This helps to get a better performance.
In the method _get_fonts_filename_windows_10_or_more, the IDWriteFontFaceReference seems to always be DWRITE_FONT_SIMULATIONS.DWRITE_FONT_SIMULATIONS_NONE, so we don't need to add a check.