[windows_fonts] Recollect
On Windows 11, if checkForUpdates is false, it will collect the font added by AddFontResource, but it won't collect the font directly added in the registry HKEY_CURRENT_USER\Software\Microsoft\Windows NT\CurrentVersion\Fonts which is what fontbase does.
If checkForUpdates is true, then it collect the font in the registry, but it doesn't collect the font added by AddFontResource.

So, we need to call the same method 2 times. Important, we need to call the method with checkForUpdates to false before checkForUpdates to true. I don't know why, but if I do the inverse, then it won't collect the font added with AddFontResource